extracted components from player
